requirements.txt
altair==4.2.0
Flask==2.1.2
ultralytics==8.3.31
opencv-python-headless==4.10.0.84
pandas>=1.3.5
streamlit>=1.28.0
protobuf==3.20.3
numpy>=1.23.5
Pillow>=9.1.0
torch>=1.8.1
torchvision>=0.9.1
PyYAML>=6.0
requests==2.27.1
scipy>=1.11.0
pymongo>=4.10.1
opencv-python